backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.2.0",
    "fastapi>=0.128.0",
    "passlib[bcrypt]==1.7.4",
    "pydantic[email]>=2.12.5",
    "pydantic-extra-types>=2.11.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "supabase>=2.27.2",
    "uvicorn>=0.40.0",
]
